% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{control}
\alias{control}
\title{Control your Spotify instance through R}
\usage{
control(action = "next")
}
\arguments{
\item{action}{the action that the user would like to take (described in description above).}
}
\description{
play, pause, jump between playing tracks, and get track info from your R console. The
control function takes a single argument, which relate to track skipping (`next`, `prev`),
stopping and starting playback (`play`, `pause`), and obtaining track information (`info`)
}
\examples{
\dontrun{
#next track
control('next')
# who sings this?
control('info')
}
}
\author{
Tad Dallas
}
\seealso{
vol
}

