% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbums.R
\name{getAlbum}
\alias{getAlbum}
\title{Get Album Info}
\usage{
getAlbum(album.id = NULL)
}
\arguments{
\item{album.id}{Character vector, Spotify album id}
}
\value{
A tibble
}
\description{
Retrieves album information. See \code{\link{parseAlbum}} for details on fields retrieved.
}
