% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbumInfo.R
\name{getAlbumInfo}
\alias{getAlbumInfo}
\title{Get additional album info}
\usage{
getAlbumInfo(album_id, token)
}
\arguments{
\item{album_id}{The Spotify ID for the album..}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
'Get an Artist’s Albums
}
\description{
function to get additional album info from an artist using the Album ID
}
