% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbuns.R
\name{getAlbuns}
\alias{getAlbuns}
\title{Get an Artist’s Albums}
\usage{
getAlbuns(id, type = "album", market = "US")
}
\arguments{
\item{id}{The Spotify ID for the artist.}

\item{market}{Optional. An ISO 3166-1 alpha-2 country code. Supply this parameter to limit the response to one particular geographical market. For example, for albums available in Sweden: market=SE.}
}
\value{
'Get an Artist’s Albums
}
\description{
function to get albuns from an artist using their ID
}

