\name{getArtistInfo}
\alias{getArtistInfo}

\title{ 

}
\description{
 
}
\usage{
getArtistInfo(artistName)
}


\arguments{
  \item{artistName}{
The name of some band you want to query. 
}
}

\value{



}

\references{
https://developer.spotify.com/
}

\author{
Tad Dallas
}



\examples{

# Nobody likes Defeater
defeater=getArtistInfo("Defeater")

# Everybody loves Kanye
kanye=getArtistInfo("Kanye West")

}

