% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbums.R
\name{getArtistAlbums}
\alias{getArtistAlbums}
\title{Get artist's albums}
\usage{
getArtistAlbums(artist.id, type = "album", market = "us")
}
\arguments{
\item{type}{Character string, defaults to 'album'. Other options are "single," "appears on," and "compilation."}

\item{market}{Character string, an ISO 3166-1 alpha-2 country code.}

\item{id}{Character string, Spotify artist id.}
}
\value{
A tibble
}
\description{
Retrieves albums for a particular \code{artist.id}. See
\code{\link{parseAlbum}} for details on fields retrieved.
}
