% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getArtistInfo.R
\name{getArtistInfo}
\alias{getArtistInfo}
\title{Get artist info}
\usage{
getArtistInfo(artistName)
}
\arguments{
\item{artistName}{The name of some band you want to query.}
}
\value{
named vector of artist characteristics
}
\description{
Returns a vector of artist information, including number of followers, the uri code, genres, and a popularity score.
}
\examples{
## Nobody loves AJJ
ajj <- getArtistInfo('AJJ')

 ## Everybody loves Kanye
kanye <- getArtistInfo("Kanye West")
}
\author{
Tad Dallas
}
\references{
https://developer.spotify.com/
}

