% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentUser.R
\name{getCurrentUser}
\alias{getCurrentUser}
\title{Get basic info of the current user}
\usage{
getCurrentUser(token)
}
\arguments{
\item{token}{An OAuth token created with \code{spotifyOAuth}.}

\item{user_id}{user id}
}
\description{
function to get basic info of the current user
}
\examples{
\dontrun{
## Example
 my_oauth <- spotifyOAuth(app_id="xxxx",client_id="yyyy",client_secret="zzzz")
 save(my_oauth, file="my_oauth")
 load("my_oauth")
 tiago <- getCurrentUser(user_id="t.mendesdantas",my_oauth)
}


}
