% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylists.R
\name{getFeaturedPlaylists}
\alias{getFeaturedPlaylists}
\title{Title}
\usage{
getFeaturedPlaylists(country = "us", limit = 20, offset = 0,
  tracks = FALSE)
}
\arguments{
\item{offset}{}
}
\description{
Title
}
