% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Get audio feature information for a single track}
\usage{
getFeatures(track_id, token)
}
\arguments{
\item{track_id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to get audio feature information for a single track identified by its unique Spotify ID.
}
