% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbums.R
\name{getNewAlbums}
\alias{getNewAlbums}
\title{Get Featured New Albums}
\usage{
getNewAlbums(country = "us", limit = 20, offset = 0)
}
\arguments{
\item{country}{Character string, an ISO 3166-1 alpha-2 country code}

\item{limit}{Numeric 1-50, number of albums to return}

\item{offset}{The index of the first item to return. Default: 0 (the first object).
Use with \code{limit} to get the next set of items.}
}
\value{
A tibble
}
\description{
Retrieve information on 'Spotify Featured Albums.' See \code{\link{parseAlbum}} for details on fields retrieved.
}
