% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylists.R
\name{getPlaylists}
\alias{getPlaylists}
\title{Get the playlists of a specific user}
\usage{
getPlaylists(user_id, offset = 0, token)
}
\arguments{
\item{user_id}{user id}

\item{offset}{The index of the first playlist to return. Default: 0 (the first object). Maximum offset: 100.000.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
get playlist from a specific user using a username
}
\description{
function to get Info about the playlists of a specific user the user_id
}
