% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylistSongs.R
\name{getPlaylistSongs}
\alias{getPlaylistSongs}
\title{Get the songs of a specific playlist}
\usage{
getPlaylistSongs(user_id, playlist_id, offset = 0, token)
}
\arguments{
\item{user_id}{User id}

\item{playlist_id}{Playlist id}

\item{offset}{The index of the first songs to return. Default: 0 (the first object). Maximum offset: 100.000.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to get songs about a specifc playlist
}
