% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylists.R
\name{getPlaylistTracks}
\alias{getPlaylistTracks}
\title{Get tracks from a playlist}
\usage{
getPlaylistTracks(user.id, playlist.id)
}
\arguments{
\item{user.id}{character string}

\item{playlist.id}{character string}
}
\description{
Get tracks from a playlist
}
