% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelated.R
\name{getRelated}
\alias{getRelated}
\title{Get Related artists}
\usage{
getRelated(artistName)
}
\arguments{
\item{artistName}{Name of the artist (can use name or Spotify ID of the artist)}
}
\description{
function to get the related artists of an Artist
}
