% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R
\name{getRelatedArtists}
\alias{getRelatedArtists}
\title{Get related artists}
\usage{
getRelatedArtists(arist.id)
}
\arguments{
\item{arist.id}{Character string of artist}
}
\description{
Get related artists
}
