% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop.R
\name{getTop}
\alias{getTop}
\title{Get top tracks of an Artist}
\usage{
getTop(id, country, token)
}
\arguments{
\item{id}{Artist ID}

\item{country}{Country of interest}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to get top tracks of an Artist by artist ID and country
}
