% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R
\name{getTopTracks}
\alias{getTopTracks}
\title{Get top tracks of an artist}
\usage{
getTopTracks(artist.id, country = "us")
}
\arguments{
\item{artist.id}{}

\item{country}{Country of interest}
}
\description{
function to get top tracks of an Artist by artist ID and country
}
