% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrack.R, R/getTracks.R
\name{getTrack}
\alias{getTrack}
\alias{getTrack}
\title{Get a Track}
\usage{
getTrack(track.id = NULL)

getTrack(track.id = NULL)
}
\arguments{
\item{track.id}{Character vector, Spotify track id. Max 20.}

\item{id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
'Get Spotify catalog information for a single track identified by its unique Spotify ID.

A tibble
}
\description{
Get Spotify catalog information for a single track identified by its unique Spotify ID.

Retrieves tracks information. See \link{\code{parseFullTrack}} for details on frields retrieved.
}
