% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTracks.R
\name{getTrackFeatures}
\alias{getTrackFeatures}
\title{Get track features}
\usage{
getTrackFeatures(track.id = NULL)
}
\arguments{
\item{track.id}{Character vector, spotify track id. Max 20.}
}
\value{
A tibble
}
\description{
Get track features
}
