% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserTop.R
\name{getUserTop}
\alias{getUserTop}
\title{Get a user's recently played tracks}
\usage{
getUserTop(user_id, artists = T, token)
}
\arguments{
\item{artists}{Whether the information requested is artists or tracks, either TRUE (artists) or FALSE (tracks)}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
Get the current user's top artists or tracks
}
\description{
Get a user's recently played tracks
}
