% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbums.R
\name{parseAlbum}
\alias{parseAlbum}
\title{Parse album object}
\usage{
parseAlbum(album_obj)
}
\arguments{
\item{album_obj}{A json-like list, album object returned by Spotify API}
}
\value{
A tibble
}
\description{
Helper function to parse useful fields from the album object. Mostly used
internally.
}
\details{
\itemize{
  \item artist name
  \item markets album is available in
  \item album name
  \item album's spotify url
  \item label album was released under
  \item popularity score, 0-100, generated by Spotify
  \item release date
  \item genres
  \item album id
}
}
