% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R
\name{parseArtist}
\alias{parseArtist}
\title{Parse artist object}
\usage{
parseArtist(art_obj)
}
\arguments{
\item{art_obj}{A json-like list, artist object returned by Spotify API}
}
\description{
Helper function to parse useful fields from the artist object.
}
\details{
\itemize{
  \item artist name
  \item arist popularity score, 0-100, generated by Spotify
  \item genres
  \item total followers
  \item artist id
  \item popularity score, 0-100, generated by Spotify
  \item release date
  \item genres
  \item album id
}
}
