% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylists.R
\name{parsePlaylist}
\alias{parsePlaylist}
\title{Parse playlist object}
\usage{
parsePlaylist(pl_obj)
}
\arguments{
\item{pl_obj}{A json-like list, playlist object returned by Spotify API}
}
\description{
Helper function to parse useful fields from the playlist object. Mostly used
internally.
}
