% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTracks.R
\name{parseShortTrack}
\alias{parseShortTrack}
\title{Parse simplified track object}
\usage{
parseShortTrack(track_obj)
}
\arguments{
\item{track_obj}{A json-like list, simplified track object returned by Spotify API}
}
\value{
A tibble
}
\description{
Helper function to parse useful fields from a simplified track obect. Mostly used
internally.
}
