% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSpotify-package.R
\docType{package}
\name{rSpotify-package}
\alias{rSpotify}
\alias{rSpotify-package}
\title{Control Spotify from R, and programmatically access related artists}
\description{
Control a Spotify instance from an R session. 
This package is a silly wrapper to a bash script 
(https://gist.github.com/wandernauta/6800547) allowing control
of Spotify functionality. It's also a bit of Spotify API searching to get the
related artists, artist info, etc. It's all in good fun. 
Feel free to contribute.
}
\author{
Tad Dallas \email{tdallas@ucdavis.edu}
}

