% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{searchAlbums}
\alias{searchAlbums}
\title{Search for albums}
\usage{
searchAlbums(search, market = "US", limit = 20, offset = 0)
}
\arguments{
\item{search}{search string}

\item{market}{market of interest, defaults to US}

\item{limit}{integer number of records to return, 1-50}

\item{offset}{integer records to offset results by, 0-100,00. Use to page through results.}
}
\value{
tibble of album info
}
\description{
Function for searching for albums by album name.
}
\seealso{
Other search functions: \code{\link{searchArtists}},
  \code{\link{searchPlaylist}}
}
