% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchArtist.R
\name{searchArtist}
\alias{searchArtist}
\title{Search for an specific Artist


function to get basic info about a searched Artist}
\usage{
searchArtist(artistName)
}
\arguments{
\item{artistName}{Name of the artist}
}
\examples{
\dontrun{
## Example
search<-searchArtist("Foo+fighters")

}


}

