% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{searchArtists}
\alias{searchArtists}
\title{Search for artists}
\usage{
searchArtists(search, market = "US", limit = 20, offset = 0)
}
\arguments{
\item{offset}{}
}
\description{
Search for artists
}
\seealso{
Other search functions: \code{\link{searchAlbums}},
  \code{\link{searchPlaylist}}
}
