% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{searchPlaylist}
\alias{searchPlaylist}
\title{Search for playlists}
\usage{
searchPlaylist(search, market = "US", limit = 20, offset = 0)
}
\arguments{
\item{offset}{}
}
\description{
Search for playlists
}
