% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTrack.R
\name{searchTrack}
\alias{searchTrack}
\title{Search for an specific Track


function to get basic info about a searched Track}
\usage{
searchTrack(trackName)
}
\arguments{
\item{trackName}{Name of the track}
}
\examples{
\dontrun{
## Example
search<-searchTrack("Foo+fighters")

}


}
