% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyOAuth.R
\name{spotifyConnect}
\alias{spotifyConnect}
\title{Add spotify token to environment}
\usage{
spotifyConnect(app_id, client_id, client_secret)
}
\arguments{
\item{client_secret}{}
}
\description{
Add spotify token to environment
}
