% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vol.R
\name{vol}
\alias{vol}
\title{Control volume from R}
\usage{
vol(mag = 1.05)
}
\arguments{
\item{mag}{factor by which to increase or decrease the speaker volume by}
}
\value{
nothing
}
\description{
This function alters the computer speaker volume, by making a call to amixer. The value `mag` is the percent change you'd like,
so a 5% decrease (increase) in volume would correspond to a mag = 0.95 (1.05).
}
\examples{
\dontrun{
#pump up the jams!!
vol(2)
#mute your garbage music when your advisor walks in
vol(0)
}
}
\author{
Tad Dallas
}

