% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbum.R
\name{getAlbum}
\alias{getAlbum}
\title{Get an Album’s Tracks}
\usage{
getAlbum(id, token)
}
\arguments{
\item{id}{The Spotify ID for the album.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
get Album from an artist using their ID
}
\description{
Get Spotify catalog information about an album’s tracks.
}
