% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbumInfo.R
\name{getAlbumInfo}
\alias{getAlbumInfo}
\title{Get additional album info}
\usage{
getAlbumInfo(id, token)
}
\arguments{
\item{id}{The Spotify ID for the artist.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}

\item{market}{Optional. An ISO 3166-1 alpha-2 country code. Supply this parameter to limit the response to one particular geographical market. For example, for albums available in Sweden: market=SE.}
}
\value{
'Get an Artist’s Albums
}
\description{
function to get additional album info from an artist using their ID
}
