% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R, R/getArtist2.R
\name{getArtist}
\alias{getArtist}
\alias{getArtist}
\title{Get Spotify catalog information for a single artist}
\usage{
getArtist(id)

getArtist(id)
}
\arguments{
\item{id}{The Spotify ID for the artist.}

\item{id}{The Spotify ID for the artist.}
}
\value{
Get Spotify catalog information for a single track identified by its unique Spotify ID.

Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
\description{
Function to get Spotify catalog information for a single artist identified by their unique Spotify ID.

Function to get Spotify catalog information for a single artist identified by their unique Spotify ID.
}
\examples{
\dontrun{
## Example


foofighters<-getArtist("7jy3rLJdDQY21OgRLCZ9sD")

}

\dontrun{
## Example


foofighters<-getArtist("7jy3rLJdDQY21OgRLCZ9sD")

}

}
