% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelated.R
\name{getRelated}
\alias{getRelated}
\title{Get Related artists


function to get the related artists of an Artist}
\usage{
getRelated(artistName, token)
}
\arguments{
\item{artistName}{Name of the artist (can use name or Spotify ID of the artist)}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
