% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTracks.R
\name{getTopTracks}
\alias{getTopTracks}
\title{Get top tracks of an Artist}
\usage{
getTopTracks(artist_id, country, token)
}
\arguments{
\item{artist_id}{Artist ID}

\item{country}{Country of interest}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to get top tracks of an Artist.
}
