% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAudioFeatures.R, R/getTrack.R
\name{getTrack}
\alias{getTrack}
\alias{getTrack}
\title{Get a Track}
\usage{
getTrack(id, token)

getTrack(id, token)
}
\arguments{
\item{id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}

\item{id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
'Get Spotify audio features analysis for a single track identified by its unique Spotify ID.

'Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
\description{
Get Spotify catalog information for a single track or a set of tracks (max 100) identified by their unique Spotify IDs.

Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
