% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrack.R
\name{getTrack}
\alias{getTrack}
\title{Get a Track}
\usage{
getTrack(track_id, token)
}
\arguments{
\item{track_id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
