% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchArtist.R
\name{searchArtist}
\alias{searchArtist}
\title{Search for a specific artist}
\usage{
searchArtist(artist_name, token)
}
\arguments{
\item{artist_name}{Name of the artist}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to get basic info about a searched artist
}
\examples{
\dontrun{

my_oauth <- spotifyOAuth(app_id="xxxx",client_id="yyyy",client_secret="zzzz")
save(my_oauth, file="my_oauth")
load("my_oauth")
foofighters <- searchArtist("Foo+fighters",token=my_oauth)
}
}
