% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{searchPlaylist}
\alias{searchPlaylist}
\title{Search for playlists}
\usage{
searchPlaylist(search, market = "US", limit = 20, offset = 0,
  tracks = FALSE)
}
\arguments{
\item{offset}{}
}
\description{
Search for playlists
}
\seealso{
Other search functions: \code{\link{searchAlbums}},
  \code{\link{searchArtists}}
}
