% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPlaylist.R
\name{searchPlaylist}
\alias{searchPlaylist}
\title{Search for playlists given a name}
\usage{
searchPlaylist(playlistName, offset = 0, token = token)
}
\arguments{
\item{playlistName}{Name of the playlist being searched for}

\item{offset}{The index of the first Playlist to return. Default: 0 (the first object). Maximum offset: 100.000.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
function to search for playlists given a name
}
