% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTrack.R
\name{searchTrack}
\alias{searchTrack}
\title{Search for a specific Track


function to get basic info about a searched Artist}
\usage{
searchTrack(track, token)
}
\arguments{
\item{track}{Name of the trach}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\examples{
\dontrun{
## Example
## my_oauth <- spotifyOAuth(app_id="xxxx",client_id="yyyy",client_secret="zzzz")
## save(my_oauth, file="my_oauth")
## load("my_oauth")
## track <- searchTrack("generator",token=my_oauth)
}
}
